Improve contact assignment  in task edit
The Assigned to function in the edit Template works now introduced the initales into the circle and the full name with the checkbox in it. LoadContactsforDropdown in edit gets the contacts from firebase and puts it in the innerHTML getConrcatForDropwdown ist the filter and template function for this function and get initales does the work for the intiales and will be executed in getContactCardForDropdown.