Subtask Counter added and saveSubtask function finished.
The saveSubtask function looks for checked checkboxes and patches it into firebase. This function is finished. The subtaskCounter funcion counters the checkbox that are checked and manipulates the innerHTML of the small task templates. Will intergrate it into the renderFunction so the right number will show up when the site is loaded.