Filter Function showOldestUrgentDueDate() added executed in init.
The showOldestUrgentDueDate() gets the object with the urgent Tasks and will create an array of the dates parses the the elements into  dates and sorts the dates in the new array from the oldest to the newest date. Then the innerHTML will be manipulated and the first urgent task Date shows up.