Introduced the functions for svg icon change, input type is changed to password and positioning of the svgs are changed and hover effect for arrow added.
The passwords are not visible at the beginning if we click the icon we can show the text of the input and it changes to an open eye. The closed eye gets the class display-none and the to logos toggle the class. Positioning of the svg are done by position absoulte and the parent container has the property position relative.