Added svgs into the input field container in login.html.  Login.js has now 3 new function that shgow or hide the eye icons and changes input from password to text type.  Responsive Design of the logos adjusted and added.
Svgs are now responsive in login and signup.html. The three functions in login.js change the input icon to and closed eye. If you click the eye icon it changes with another function the closed eye to an open eye and the input type will be changed to an text type. If you click the open eye the closed eye will show up again and the type is password again. The works with display-none on the svgs.